boil

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -9,7 +9,7 @@ int _atoi(char *s)
 {
 	int result = 0;
 	int sign = 1;
-	long long i = 0;
+	int i = 0;
 
 	while (s[i] != '\0' && (s[i] < '0' || s[i] > '9'))
 	{
